Move progress graph to activity screen
