+++ app/src/main/java/com/github/k409/fitflow/ui/common/Components.kt
@@ -1,10 +1,19 @@
 package com.github.k409.fitflow.ui.common
 
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.pager.PagerState
+import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -16,6 +25,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 
 @Composable
@@ -62,6 +73,38 @@ fun FancyIndicatorTabs(
                     )
                 }
             }
+        }
+    }
+}
+
+@OptIn(ExperimentalFoundationApi::class)
+@Composable
+fun HorizontalPagerIndicator(
+    modifier: Modifier = Modifier,
+    pagerState: PagerState,
+    defaultRadius: Dp = 8.dp,
+    space: Dp = 6.dp,
+    activeColor: Color = MaterialTheme.colorScheme.onSurface,
+    inactiveColor: Color = activeColor.copy(alpha = 0.5f)
+) {
+    Row(
+        modifier
+            .wrapContentHeight()
+            .fillMaxWidth(),
+        horizontalArrangement = Arrangement.Center,
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        repeat(pagerState.pageCount) { iteration ->
+            val color =
+                if (pagerState.currentPage == iteration) activeColor else inactiveColor
+
+            Box(
+                modifier = Modifier
+                    .padding(horizontal = space / 2)
+                    .clip(CircleShape)
+                    .background(color)
+                    .size(defaultRadius)
+            )
         }
     }
 }
